Remove XSS vulnerabilities in the game.html model.
JSON is moved to HTML data-* attributes which allows them to be escaped
fully.

Also now using Django's built-in filters wherever possible (especially
in preference to doing checks within JS). Helps with readability and
reliability (eg: differentiating the string "None" and value None).
distinguishable from a variable which is None.